Move `Dates` stuff back from package extension
