+++ Project.toml
@@ -20,7 +20,6 @@ Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 
 [extensions]
-CompatDatesExt = "Dates"
 CompatLinearAlgebraExt = "LinearAlgebra"
 
 [targets]
